backend/requirements.txt
langchain
langchain-community
transformers
sentence-transformers
faiss-cpu
torch
accelerate
python-dotenv
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
langchain-huggingface
langchain-core
sqlalchemy
bcrypt
